Data-driven planet colors and respecting the world border

+++ src/main/java/dev/devce/rocketnautics/client/RenderDataSelector.java
@@ -1,5 +1,5 @@
 package dev.devce.rocketnautics.client;
 
 public interface RenderDataSelector {
-    byte select(int x, int y);
+    int select(int x, int y);
 }
